+++ packages/gatsby-transformer-cloudinary/gatsby-node/create-asset-nodes-from-data.js
@@ -1,3 +1,7 @@
+const flatMap = require('lodash/flatMap');
+const get = require('lodash/get');
+const set = require('lodash/set');
+const unset = require('lodash/unset');
 const { createImageNode } = require('../create-image-node');
 
 exports.createAssetNodesFromData = ({
@@ -6,11 +10,14 @@ exports.createAssetNodesFromData = ({
   createNodeId,
   createContentDigest,
 }) => {
-  const assetDataKeys = getAssetDataKeys(node);
-
-  assetDataKeys.forEach(assetDataKey => {
-    const assetData = { ...node[assetDataKey] };
-    delete node[assetDataKey];
+  const assetDataPaths = getAssetDataPaths({ node });
+  assetDataPaths.forEach(assetDataPath => {
+    const assetData = {
+      ...get(node, assetDataPath),
+    };
+    unset(node, assetDataPath);
+    const assetDataPathParts = assetDataPath.split('.');
+    const relationshipName = assetDataPathParts[assetDataPathParts.length - 1];
     if (verifyAssetData(assetData)) {
       createCloudinaryAssetNode({
         assetData,
@@ -18,7 +25,8 @@ exports.createAssetNodesFromData = ({
         createNode,
         createNodeId,
         parentNode: node,
-        relationshipName: assetDataKey,
+        relationshipName,
+        assetDataPath,
       });
     }
   });
@@ -41,8 +49,40 @@ function getAssetDataKeys(node) {
   });
 }
 
+function getAssetDataPaths({ node, basePath = '' }) {
+  const currentNode = basePath === '' ? node : get(node, basePath);
+
+  const directAssetDataPaths = Object.keys(currentNode)
+    .filter(key => {
+      return currentNode[key] && currentNode[key].cloudinaryAssetData === true;
+    })
+    .map(subPath => {
+      return basePath === '' ? subPath : `${basePath}.${subPath}`;
+    });
+
+  const objectPaths = Object.keys(currentNode)
+    .filter(key => {
+      return isObject(currentNode[key]);
+    })
+    .map(subPath => {
+      return basePath === '' ? subPath : `${basePath}.${subPath}`;
+    });
+
+  const indirectAssetDataPaths = flatMap(objectPaths, objectPath => {
+    return getAssetDataPaths({ node, basePath: objectPath });
+  });
+
+  const assetDataPaths = [...directAssetDataPaths, ...indirectAssetDataPaths];
+  return assetDataPaths;
+}
+
+function isObject(thing) {
+  return typeof thing === 'object' && thing != null;
+}
+
 function createCloudinaryAssetNode({
   assetData: { cloudName, originalHeight, originalWidth, publicId, version },
+  assetDataPath = null,
   createContentDigest,
   createNode,
   createNodeId,
@@ -68,6 +108,6 @@ function createCloudinaryAssetNode({
   createNode(imageNode, { name: 'gatsby-transformer-cloudinary' });
 
   // Tell Gatsby to add `${relationshipName}` to the parent node.
-  const relationshipKey = `${relationshipName}___NODE`;
-  parentNode[relationshipKey] = imageNode.id;
+  const relationshipKey = `${assetDataPath || relationshipName}___NODE`;
+  set(parentNode, relationshipKey, imageNode.id);
 }