+++ packages/gatsby-transformer-cloudinary/CHANGELOG.md
@@ -1,3 +1,9 @@
+# Version next
+
+Additions:
+
+- Added the ability to use both width and height parameters simultaneously for fixed queries.
+
 # Version 2.0.0
 
 Changes: