+++ packages/gatsby-transformer-cloudinary/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Version next
 
+Fixes:
+
+- Deeply nested asset data is now transformed into CloudinaryAsset nodes.
+
 Additions:
 
 - Added the ability to use both width and height parameters simultaneously for fixed queries.