+++ packages/gatsby-transformer-cloudinary/CHANGELOG.md
@@ -4,6 +4,10 @@ Additions:
 
 - Added logging for each time we have to fetch a base64 image from Cloudinary to explain long query steps in the Gatsby build process.
 
+Fixes:
+
+- Fluid images use defaultBase64 images when they are provided.
+
 # Version 2.1.0
 
 Additions: